consolidate theming utils
