+++ .changeset/tidy-themes-consolidate.md
@@ -1,0 +1,9 @@
+---
+"@fluentui-react-native/design": patch
+"@fluentui-react-native/theme-tokens": patch
+"@fluentui-react-native/themed-stylesheet": patch
+"@fluentui-react-native/theming-utils": patch
+"@fluentui-react-native/tokens": patch
+---
+
+Consolidate legacy theming and styling implementations into design submodules and deprecate their former packages as compatibility shims.
