+++ packages/agentic/design/package.json
@@ -26,6 +26,12 @@
       "import": "./lib/tokens/global.generated.js",
       "default": "./src/tokens/global.generated.ts"
     },
+    "./tokens/legacy": {
+      "types": "./lib/tokens/legacy/index.d.ts",
+      "react-native": "./src/tokens/legacy/index.ts",
+      "import": "./lib/tokens/legacy/index.js",
+      "default": "./src/tokens/legacy/index.ts"
+    },
     "./theming": {
       "types": "./lib/theming/index.d.ts",
       "react-native": "./src/theming/index.ts",
@@ -44,6 +50,18 @@
       "import": "./lib/styling/index.js",
       "default": "./src/styling/index.ts"
     },
+    "./styling/stylesheet": {
+      "types": "./lib/styling/stylesheet/index.d.ts",
+      "react-native": "./src/styling/stylesheet/index.ts",
+      "import": "./lib/styling/stylesheet/index.js",
+      "default": "./src/styling/stylesheet/index.ts"
+    },
+    "./styling/tokens": {
+      "types": "./lib/styling/tokens/index.d.ts",
+      "react-native": "./src/styling/tokens/index.ts",
+      "import": "./lib/styling/tokens/index.js",
+      "default": "./src/styling/tokens/index.ts"
+    },
     "./package.json": "./package.json"
   },
   "scripts": {
@@ -56,15 +74,17 @@
     "update-snapshots": "fluentui-scripts jest -u"
   },
   "dependencies": {
-    "@fluentui-react-native/framework-base": "workspace:*"
+    "@fluentui-react-native/adapters": "workspace:*",
+    "@fluentui-react-native/design-tokens-android": "catalog:",
+    "@fluentui-react-native/design-tokens-ios": "catalog:",
+    "@fluentui-react-native/design-tokens-win32": "catalog:",
+    "@fluentui-react-native/design-tokens-windows": "catalog:",
+    "@fluentui-react-native/framework-base": "workspace:*",
+    "assert-never": "^1.2.1"
   },
   "devDependencies": {
     "@babel/core": "catalog:",
-    "@fluentui-react-native/design-tokens-android": "catalog:",
-    "@fluentui-react-native/design-tokens-ios": "catalog:",
     "@fluentui-react-native/design-tokens-macos": "catalog:",
-    "@fluentui-react-native/design-tokens-win32": "catalog:",
-    "@fluentui-react-native/design-tokens-windows": "catalog:",
     "@fluentui-react-native/scripts": "workspace:*",
     "@office-iss/react-native-win32": "^0.81.0",
     "@react-native-community/cli": "^20.0.0",
