+++ packages/agentic/design/src/styling/stylesheet/index.ts
@@ -1,0 +1,2 @@
+export { themedStyleSheet } from './themedStyleSheet';
+export type { NamedStyles } from './themedStyleSheet';
