+++ packages/agentic/design/src/styling/tokens/color-tokens.ts
@@ -1,6 +1,6 @@
 import type { ColorValue } from 'react-native';
 
-import type { Theme, ThemeColorDefinition } from '@fluentui-react-native/design/theming';
+import type { Theme, ThemeColorDefinition } from '../../theming';
 
 import type { OperationSet } from './token.types';
