+++ packages/agentic/design/src/styling/tokens/text-tokens.ts
@@ -1,7 +1,7 @@
 import type { TextProps, TextStyle } from 'react-native';
 
 import type { ITextProps, ITextStyle } from '@fluentui-react-native/adapters';
-import type { Theme, Typography } from '@fluentui-react-native/design/theming';
+import type { Theme, Typography } from '../../theming';
 
 import { styleFunction } from './token.function';
 import type { TokenBuilder } from './tokenBuilder';
