+++ packages/agentic/design/src/theming/index.ts
@@ -35,3 +35,5 @@ export type { ThemeProviderProps } from './ThemeProvider';
 export { ThemeReference } from './themeReference';
 export type { OnThemeChange, ThemeRecipe, ThemeTransform } from './themeReference';
 export { getCurrentAppearance, isHighContrast, setIsHighContrast } from './platformUtils';
+export { mapFontPipelineToTheme, mapPipelineToTheme } from './mapPipelineToTheme';
+export { mapPipelineToShadow } from './mapPipelineToShadow';
