+++ packages/agentic/design/src/theming/mapPipelineToTheme.android.ts
@@ -1,4 +1,5 @@
-import type { AliasColorTokens, Variants, VariantValue } from '@fluentui-react-native/design/theming';
+import type { AliasColorTokens } from './types/Color.types';
+import type { Variants, VariantValue } from './types/Typography.types';
 
 // API that maps tokens coming from the android token pipeline to Theme color values.
 export function mapPipelineToTheme(pipelineOutput: any): AliasColorTokens {
