+++ packages/agentic/design/src/tokens/legacy/index.ts
@@ -1,0 +1,2 @@
+export { default as globalTokens } from './tokens-global';
+export { getAliasTokens, getShadowTokens } from './getTokens';
