+++ packages/agentic/design/src/tokens/legacy/tokens-global.win32.ts
