+++ packages/agentic/design/tsconfig.json
@@ -11,6 +11,9 @@
   "exclude": ["**/*.stories.ts", "**/*.stories.tsx"],
   "references": [
     {
+      "path": "../../utils/adapters/tsconfig.json"
+    },
+    {
       "path": "../../framework-base/tsconfig.json"
     },
     {
