+++ packages/framework/themed-stylesheet/src/index.ts
@@ -1,2 +1,5 @@
-export { themedStyleSheet } from './themedStyleSheet';
-export type { NamedStyles } from './themedStyleSheet';
+// This package is a compatibility shim. The legacy themed StyleSheet helper now
+// lives in the `@fluentui-react-native/design/styling/stylesheet` submodule.
+// These explicit re-exports preserve the existing package entry point.
+export { themedStyleSheet } from '@fluentui-react-native/design/styling/stylesheet';
+export type { NamedStyles } from '@fluentui-react-native/design/styling/stylesheet';
