+++ packages/theming/theme-tokens/README.md
@@ -1,3 +1,7 @@
 # Theme-tokens
 
-Defines values for tokens used to fill out themes, such as color, typography, and spacing values.
+> **Deprecated:** Import `globalTokens`, `getAliasTokens`, and `getShadowTokens`
+> from `@fluentui-react-native/design/tokens/legacy` instead.
+
+This package is a compatibility shim and will not receive new APIs. It preserves
+the existing entry point while consumers migrate to the design package.
