+++ packages/theming/theme-tokens/src/index.ts
@@ -1,2 +1,4 @@
-export { default as globalTokens } from './tokens-global';
-export { getAliasTokens, getShadowTokens } from './getTokens';
+// This package is a compatibility shim. Legacy Fluent token values now live in
+// the `@fluentui-react-native/design/tokens/legacy` submodule. These explicit
+// re-exports preserve the existing `@fluentui-react-native/theme-tokens` entry point.
+export { getAliasTokens, getShadowTokens, globalTokens } from '@fluentui-react-native/design/tokens/legacy';
