+++ packages/theming/theming-utils/README.md
@@ -1,7 +1,7 @@
-# Tokens
+# Theming Utils
 
-> **Deprecated:** Import these token-to-style helpers from
-> `@fluentui-react-native/design/styling/tokens` instead.
+> **Deprecated:** Import appearance and token-pipeline helpers from
+> `@fluentui-react-native/design/theming` instead.
 
 This package is a compatibility shim and will not receive new APIs. It preserves
 the existing entry point while consumers migrate to the design package.
