+++ packages/theming/theming-utils/src/index.ts
@@ -1,3 +1,11 @@
-export { getCurrentAppearance, isHighContrast, setIsHighContrast } from '@fluentui-react-native/design/theming';
-export { mapPipelineToTheme, mapFontPipelineToTheme } from './mapPipelineToTheme';
-export { mapPipelineToShadow } from './mapPipelineToShadow';
+// This package is a compatibility shim. Theme appearance and token-pipeline
+// utilities now live in the `@fluentui-react-native/design/theming` submodule.
+// These explicit re-exports preserve the existing package entry point.
+export {
+  getCurrentAppearance,
+  isHighContrast,
+  mapFontPipelineToTheme,
+  mapPipelineToShadow,
+  mapPipelineToTheme,
+  setIsHighContrast,
+} from '@fluentui-react-native/design/theming';
