+++ packages/theming/theming-utils/tsconfig.json
@@ -13,9 +13,6 @@
     },
     {
       "path": "../../../scripts/tsconfig.json"
-    },
-    {
-      "path": "../../utils/tokens/tsconfig.json"
     }
   ]
 }
