+++ packages/utils/tokens/package.json
@@ -30,7 +30,6 @@
     "update-snapshots": "fluentui-scripts jest -u"
   },
   "dependencies": {
-    "@fluentui-react-native/adapters": "workspace:*",
     "@fluentui-react-native/design": "workspace:*"
   },
   "devDependencies": {
