+++ packages/utils/tokens/src/index.ts
@@ -1,14 +1,28 @@
-export { borderStyles, borderTokens } from './border-tokens';
-export type { IBorderTokens } from './border-tokens';
-export { backgroundColorTokens, colorTokens, foregroundColorTokens, getPaletteFromTheme } from './color-tokens';
-export type { IBackgroundColorTokens, IColorTokens, IForegroundColorTokens } from './color-tokens';
-export { fontStyles, textTokens } from './text-tokens';
-export type { FontDecorationTokens, FontStyleTokens, FontTokens, FontVariantTokens } from './text-tokens';
-export { layoutStyles, layoutTokens } from './layout-tokens';
-export type { LayoutTokens } from './layout-tokens';
-export { shadowStyles, shadowTokens } from './shadow-tokens';
-export type { IShadowTokens } from './shadow-tokens';
-export { tokenBuilder } from './tokenBuilder';
-export type { TokenBuilder } from './tokenBuilder';
-export type { LookupThemePart, OperationSet, StyleFactoryOperation, StyleFactoryFunction, StyleFactoryFunctionRaw } from './token.types';
-export { styleFunction } from './token.function';
+// This package is a compatibility shim. Legacy token-to-style helpers now live
+// in the `@fluentui-react-native/design/styling/tokens` submodule. These explicit
+// re-exports preserve the existing `@fluentui-react-native/tokens` entry point.
+export { borderStyles, borderTokens } from '@fluentui-react-native/design/styling/tokens';
+export type { IBorderTokens } from '@fluentui-react-native/design/styling/tokens';
+export {
+  backgroundColorTokens,
+  colorTokens,
+  foregroundColorTokens,
+  getPaletteFromTheme,
+} from '@fluentui-react-native/design/styling/tokens';
+export type { IBackgroundColorTokens, IColorTokens, IForegroundColorTokens } from '@fluentui-react-native/design/styling/tokens';
+export { fontStyles, textTokens } from '@fluentui-react-native/design/styling/tokens';
+export type { FontDecorationTokens, FontStyleTokens, FontTokens, FontVariantTokens } from '@fluentui-react-native/design/styling/tokens';
+export { layoutStyles, layoutTokens } from '@fluentui-react-native/design/styling/tokens';
+export type { LayoutTokens } from '@fluentui-react-native/design/styling/tokens';
+export { shadowStyles, shadowTokens } from '@fluentui-react-native/design/styling/tokens';
+export type { IShadowTokens } from '@fluentui-react-native/design/styling/tokens';
+export { tokenBuilder } from '@fluentui-react-native/design/styling/tokens';
+export type { TokenBuilder } from '@fluentui-react-native/design/styling/tokens';
+export type {
+  LookupThemePart,
+  OperationSet,
+  StyleFactoryFunction,
+  StyleFactoryFunctionRaw,
+  StyleFactoryOperation,
+} from '@fluentui-react-native/design/styling/tokens';
+export { styleFunction } from '@fluentui-react-native/design/styling/tokens';
