+++ packages/utils/tokens/tsconfig.json
@@ -9,9 +9,6 @@
   "include": ["src"],
   "references": [
     {
-      "path": "../adapters/tsconfig.json"
-    },
-    {
       "path": "../../agentic/design/tsconfig.json"
     },
     {
